fix: Use given timeout value for failure reason
My fix for issue #33 introduced a timeout failure reason as argument for
the `on_fail` callback. This reason was intended to contain the given
timeout value.

The timeout value is changed to contain the current timestamp rendering
this value useless for the message.

This patch introduces another variable storing the given value and uses
this new value for the reason.